Fixed the word8 overflow for indices for FDir filters which can be very high.

+++ Dragonet/c_impl/dpdkDrv/filterMgmt.c
@@ -27,7 +27,7 @@ bool e10k_ctrl_fdir_unset(struct state *state, uint8_t index)
 
 
 bool e10k_ctrl_fdir_set(struct state *state,
-        uint16_t index, uint8_t queue,
+        uint32_t index, uint8_t queue,
         uint32_t src_ip, uint32_t dst_ip, uint16_t src_port, uint16_t dst_port,
         uint16_t l4_type, uint16_t mask)
 {
